requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.10
django-filter==1.0.2
djangorestframework==3.6.2
gunicorn==19.7.1
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyparsing==2.2.0
six==1.10.0
